api/requirements.txt
fastapi==0.129.0
uvicorn==0.41.0
supabase==2.28.0
python-dotenv==1.2.1
pydantic==2.12.5
